# docs/swagger/openapi.yaml
openapi: 3.1.0
info:
  title: LeetGaming Pro API
  description: |
    # 🎮 LeetGaming Pro - Competitive Gaming Platform API

    Welcome to the LeetGaming Pro API documentation. This API powers a professional-grade
    competitive gaming platform with features including:

    - **Authentication & Security**: OAuth2, MFA/TOTP, JWT tokens
    - **Matchmaking**: Glicko-2 rating system, skill-based matching
    - **Tournaments**: Bracket generation, prize pools, blockchain verification
    - **Replays**: Upload, analysis, sharing with demo parsing
    - **Squads**: Team management, invitations, statistics
    - **Billing**: Subscriptions, withdrawals, audit trails

    ## 🔐 Authentication

    Most endpoints require authentication via Bearer token:
    ```
    Authorization: Bearer <your-jwt-token>
    ```

    ## 📊 Rate Limits

    | Tier | Requests/min | Burst |
    |------|--------------|-------|
    | Free | 60 | 10 |
    | Pro | 300 | 50 |
    | Enterprise | 1000 | 200 |

    ## 🌐 Base URLs

    | Environment | URL |
    |-------------|-----|
    | Production | https://api.leetgaming.gg |
    | Staging | https://api.staging.leetgaming.gg |
    | Development | http://localhost:8080 |

  version: 2.0.0
  contact:
    name: LeetGaming Pro Support
    url: https://leetgaming.gg/support
    email: api@leetgaming.gg
  license:
    name: Proprietary
    url: https://leetgaming.gg/terms
servers:
  - url: https://api.leetgaming.gg
    description: Production
...
    description: Staging
...
paths:
...
  /players/{player_id}/rating:
    get:
      tags: [Players]
      summary: Get player rating (Glicko-2)
      operationId: getPlayerRating
      parameters:
        - $ref: "#/components/parameters/PlayerId"
      responses:
        "200":
          description: Player rating retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerRating"

  # Matchmaking
...
  /leaderboard:
    get:
      tags: [Leaderboard]
      summary: Get global leaderboard
      operationId: getLeaderboard
      parameters:
        - name: game_id
          in: query
          schema:
            type: string
        - name: region
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
            maximum: 500
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Leaderboard data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaderboardResponse"

  # Tournaments
  /tournaments:
    get:
      tags: [Tournaments]
      summary: List tournaments
      operationId: listTournaments
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [upcoming, active, completed]
        - name: game_id
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Tournament list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tournament"

    post:
      tags: [Tournaments]
      summary: Create a tournament
      operationId: createTournament
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTournamentRequest"
      responses:
        "201":
          description: Tournament created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tournament"
...
  /ranks/distribution:
    get:
      tags: [Leaderboard]
      summary: List Distribution
      operationId: get_ranks_distribution
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /tournaments/upcoming:
    get:
      tags: [Tournaments]
      summary: List Upcoming
      operationId: get_tournaments_upcoming
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /tournaments/{id}:
    get:
      tags: [Tournaments]
      summary: Get Tournaments by ID
      operationId: get_tournaments_id
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /players/{player_id}/tournaments:
    get:
      tags: [Players]
      summary: Get Tournaments by ID
      operationId: get_players_player_id_tournaments
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /organizers/{organizer_id}/tournaments:
    get:
      tags: [Tournaments]
      summary: Get Tournaments by ID
      operationId: get_organizers_organizer_id_tournaments
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PlayerId:
      name: player_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      required: [message, code]
      properties:
        message:
          type: string
        code:
          type: string
        details:
          type: object
...
    PlayerRating:
      type: object
      properties:
        player_id:
          type: string
          format: uuid
        rating:
          type: number
          format: double
        rating_deviation:
          type: number
          format: double
        volatility:
          type: number
          format: double
        rank:
          type: string
          enum:
            [
              Unranked,
              Silver,
              Gold,
              Platinum,
              Diamond,
              Master,
              Grandmaster,
              Global Elite,
            ]
        games_played:
          type: integer
        wins:
          type: integer
        losses:
          type: integer
...
    LeaderboardResponse:
      type: object
      properties:
        entries:
          type: array
          items:
            $ref: "#/components/schemas/LeaderboardEntry"
        total_count:
          type: integer
    LeaderboardEntry:
      type: object
      properties:
        rank:
          type: integer
        player_id:
          type: string
          format: uuid
        username:
          type: string
        rating:
          type: number
        tier:
          type: string
        games_played:
          type: integer
        win_rate:
          type: number
    Tournament:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        game_id:
          type: string
        status:
          type: string
          enum: [draft, registration, active, completed, cancelled]
        format:
          type: string
          enum: [single_elimination, double_elimination, round_robin, swiss]
        max_participants:
          type: integer
        prize_pool_amount:
          type: string
        prize_pool_currency:
          type: string
        starts_at:
          type: string
          format: date-time
    CreateTournamentRequest:
      type: object
      required: [name, game_id, format, max_participants]
      properties:
        name:
          type: string
        game_id:
          type: string
        format:
          type: string
          enum: [single_elimination, double_elimination, round_robin, swiss]
        max_participants:
          type: integer
          minimum: 4
          maximum: 1024
        entry_fee:
          type: number
        starts_at:
          type: string
          format: date-time
...
security:
  - bearerAuth: []